adding type checks to public API in builders.
