+++ src/bloqade/builder/pragmas.py
@@ -1,19 +1,18 @@
 from bloqade.builder.base import Builder
 from bloqade.builder.typing import ScalarType
+from bloqade.builder.parallelize import Parallelize
+from bloqade.builder.flatten import Flatten
+
 from beartype.typing import List, TYPE_CHECKING
 from beartype import beartype
 
 if TYPE_CHECKING:
-    from bloqade.builder.parallelize import Parallelize
     from bloqade.builder.assign import Assign, BatchAssign
-    from bloqade.builder.flatten import Flatten
 
 
 class Flattenable(Builder):
     @beartype
-    def flatten(self, orders: List[str]) -> "Flatten":
-        from bloqade.builder.flatten import Flatten
-
+    def flatten(self, orders: List[str]) -> Flatten:
         return Flatten(orders, self)
 
 
@@ -53,7 +52,7 @@ class BatchAssignable(Builder):
 
 class Parallelizable(Builder):
     @beartype
-    def parallelize(self, cluster_spacing: ScalarType) -> "Parallelize":
+    def parallelize(self, cluster_spacing: ScalarType) -> Parallelize:
         """
         Parallelize the current problem (register & sequnece) to fill entire FOV
         with the given cluster spacing.
@@ -73,6 +72,4 @@ class Parallelizable(Builder):
             >>> prob = prob.parallelize(7.2)
 
         """
-        from bloqade.builder.parallelize import Parallelize
-
         return Parallelize(cluster_spacing, self)
