+++ src/bloqade/builder/sequence_builder.py
@@ -3,9 +3,9 @@ from bloqade.builder.route import PragmaRoute
 from bloqade.builder.base import Builder
 
 
-class SequenceBuilder(PragmaRoute):
-    __match_args__ = ("sequence", "__parent__")
+class SequenceBuilder(PragmaRoute, Builder):
+    __match_args__ = ("_sequence", "__parent__")
 
     def __init__(self, sequence: Sequence, parent: Builder):
-        self.sequence = sequence
-        self.__parent__ = parent
+        super().__init__(parent)
+        self._sequence = sequence
