+++ src/bloqade/builder/typing.py
@@ -1,7 +1,7 @@
 from bloqade.ir.scalar import Scalar
 from decimal import Decimal
 from numbers import Real
-from beartype import List, Union
+from beartype.typing import List, Union
 
 
 ScalarType = Union[Real, Decimal, str, Scalar]
